Parent and Child - IPC
Create an Inter Process Communication between Parent and Child process, in which one process sends a String message to a second process, and the second process receives the message. For example, if the first process sends the message Hello, the second process receives Hello. Write a C program to illustrate this scenario. Use read(), write() system calls for sending and receiving messages.